+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  * Added support for Volatiles (#460)
  * Added JSON encoder and decoder in `basilisp.json` namespace (#484)
  * Added support for generically diffing Basilisp data structures in `basilisp.data` namespace (#555)
+ * Added support for artificially abstract bases classes in `deftype`, `defrecord`, and `reify` types (#565)
  * Added CLI argument parser in `basilisp.cli-tools` namespace (#535)
 
 ### Changed